pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repo-health-analyzer"
version = "0.1.0"
description = "A system to analyze the health of a GitHub repository over time"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your.email@example.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
  "requests>=2.25.0",
  "python-dotenv>=0.19.0",
  "click>=8.0.0",
  "rich>=10.0.0",
  "python-dateutil>=2.8.0",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
]

[dependency-groups]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "black>=22.0.0",
  "isort>=5.0.0",
  "flake8>=4.0.0",
  "mypy>=0.900",
  "responses>=0.20.0",
]

[project.scripts]
repo-health = "repo_health.__main__:main"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["src/repo_health"]
